m2+m3/______m1/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.0
telethon>=2.0.0b6 
aiosqlite==0.20.0
jinja2==3.1.4
cryptg==0.3.0